Split traits to contain implementation of single interface only.
